move update scenario tests to their own package, update pipelines to use azuresample external release, add some retryOnFailures
