+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/atpop/update/TestCase1922530.java
@@ -20,7 +20,7 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.msal.automationapp.testpass.broker.atpop;
+package com.microsoft.identity.client.msal.automationapp.testpass.broker.atpop.update;
 
 import com.microsoft.identity.client.Prompt;
 import com.microsoft.identity.client.msal.automationapp.R;
@@ -42,14 +42,15 @@ import org.junit.Test;
 
 import java.util.Arrays;
 
-// [Non-Joined] [Update-old-to-V5] Acquire PoP token Silent
-// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/1922531
-public class TestCase1922531 extends AbstractMsalBrokerUpdateTest {
+// [Joined] [Update-old-to-V5] Acquire PoP token Silent
+// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/1922530
+public class TestCase1922530  extends AbstractMsalBrokerUpdateTest {
     @Test
-    public void test_1922531() throws Throwable {
+    public void test_1922530() throws Throwable {
         final String username = mLabAccount.getUsername();
         final String password = mLabAccount.getPassword();
 
+        mBroker.performDeviceRegistration(username, password);
         final MsalSdk msalSdk = new MsalSdk();
 
         final MsalAuthTestParams authTestParams = MsalAuthTestParams.builder()
@@ -82,6 +83,9 @@ public class TestCase1922531 extends AbstractMsalBrokerUpdateTest {
         authResult.assertSuccess();
         MsalAuthResult.verifyATForPop(authResult.getAccessToken());
 
+        // Update the app
+        // TO-DO after the API to retrieve the broker version is added,
+        // we can add an assert to verify that the broker version is updated as expected
         mBroker.update();
 
         // start silent token request in MSAL
