+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/atpop/update/TestCase1922547.java
@@ -20,7 +20,8 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.msal.automationapp.testpass.broker.atpop;
+
+package com.microsoft.identity.client.msal.automationapp.testpass.broker.atpop.update;
 
 import com.microsoft.identity.client.Prompt;
 import com.microsoft.identity.client.msal.automationapp.R;
@@ -43,13 +44,15 @@ import org.junit.Test;
 
 import java.util.Arrays;
 
-// [Non-Joined] [Update-old-to-V5] Generate SHR
-// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/1922549
-public class TestCase1922549 extends AbstractMsalBrokerUpdateTest {
+// [Joined] [Update-old-to-V5] Generate SHR
+// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/1922547
+public class TestCase1922547  extends AbstractMsalBrokerUpdateTest {
     @Test
-    public void test_1922549() throws Throwable {
+    public void test_1922547() throws Throwable {
         final String username = mLabAccount.getUsername();
         final String password = mLabAccount.getPassword();
+
+        mBroker.performDeviceRegistration(username, password);
 
         final MsalSdk msalSdk = new MsalSdk();
 
@@ -84,11 +87,11 @@ public class TestCase1922549 extends AbstractMsalBrokerUpdateTest {
         MsalAuthResult.verifyATForPop(authResult.getAccessToken());
 
         mBroker.update();
-
         String shr = msalSdk.generateSHR(authTestParams, TokenRequestTimeout.SHORT);
         Assert.assertNotNull(shr);
         MsalAuthResult.verifyATForPop(shr);
     }
+
     @Override
     public LabQuery getLabQuery() {
         return LabQuery.builder()
@@ -116,3 +119,4 @@ public class TestCase1922549 extends AbstractMsalBrokerUpdateTest {
         return R.raw.msal_config_default;
     }
 }
+
